Check before detaching pages if deleted entity was soft deleted

+++ src/Traits/Pageable.php
@@ -72,7 +72,8 @@ trait Pageable
     public static function bootPageable()
     {
         static::deleted(function (self $model) {
-            $model->pages()->detach();
+            // Check if this is a soft delete or not by checking if `SoftDeletes::isForceDeleting` method exists
+            (method_exists($model, 'isForceDeleting') && ! $model->isForceDeleting()) || $model->pages()->detach();
         });
     }
